cleanup-0: refactoring graph_impl::Net moved to impl/graph_traits.hpp chromium-like indentation graph_impl::Net declaration removed from impl/meta_utils.hpp
