+++ src/graph.hpp
@@ -9,7 +9,7 @@
 namespace graph {
 
 template<typename Node, typename... Traits>
-    requires (graph_impl::Trait<Traits> && ...)
+  requires (graph_impl::Trait<Traits> && ...)
 using Graph = graph_impl::Graph<
     Node,
     graph_impl::build_graph_traits<Node, Traits...>,
